Update master.py
Found the bug in the auto shutdown. The pi does not have CMOS time so it has to get it from the NTP. I think the time that is passed in the morning when the bug is exhibited is the time from the night before when it last shut down. I have added a minimum 10 min up-time before the 10 min delay before shutdown that should stop this glitch.

+++ scripts/master.py
@@ -145,6 +145,7 @@ try:
                 
     ########Get Up Time##########
     def uptime():
+        global uptime
         uptimesec = os.popen("awk '{print $1}' /proc/uptime").readline()
         m, s = divmod(float(uptimesec), 60)
         h, m = divmod(m, 60)
@@ -650,16 +651,17 @@ try:
             hourofday = datetime.datetime.now().time()
             autoshutdowntime = datetime.time(19) 
             if safetoshutdown and hourofday >= autoshutdowntime and home and not testing and not mobile:
+                uptime()
+                if uptime > "00:00:10:00":
                     print("System will Powered off in 10 Min")
                     time.sleep(600)
                     os.system("poweroff")
                     
             time.sleep(30)
-        
-        
+            
         time.sleep(30)
 
-
+        
 finally:
     os.unlink(pidfile) 
     pass
